Important fixes: - Line position measuring that was off by one is now corrected (SCNText positioning) - *Chamfer Radius does not count as bounding box*, so account for that -- This is still buggy / wrong. The empty child is showing up pretty far away from other pieces so that's a thing. -- (empty space now has a non-zero position? I dunno. removing it and setting to nothing is also wrong) - LineIterator scrapped for nodeLength
** NOTE **
So this works right now because I happen to be iterating the 'CodeBlockItemList'.
The root has a list, the list has children, and if I find out one of the children is a ThingOfInterest,
I pass the *parent* of the inner to visit; this means that I've now passed a function / enum / class / etc to visitChildrenOf,
which means it *usually* skips the internal declaration and iterates the nodes. So what do I do?
WAIT TILL TOMORROW.